requirements.txt
yfinance
pydantic-ai
asyncio
streamlit
python-dotenv
duckduckgo-search
